Fractal flame editor: FIXED: iteration overflow on large images (or very high quality settings)

+++ src/org/jwildfire/create/tina/base/RasterPoint.java
@@ -20,5 +20,5 @@ public class RasterPoint {
   public double red;
   public double green;
   public double blue;
-  public int count;
+  public long count;
 }
